# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TeklifPro API",
    "version": "1.0.0",
    "description": "TeklifPro is a Turkish B2B proposal management platform designed to streamline the creation, management, and delivery of business proposals. The API provides comprehensive endpoints for authentication, proposal management, customer management, product catalogs, AI-powered suggestions, and integration with accounting systems via Parasut.",
    "contact": {
      "name": "Sercan Hosgoerns",
      "email": "hosgorens@gmail.com"
    },
    "license": {
      "name": "All Rights Reserved"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.teklifpro.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/verify-email": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request email verification",
        "description": "Sends a verification email to the user. Called after registration to verify email address. | Kullanıcıya doğrulama e-postası gönderir.",
        "operationId": "sendVerificationEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "sercan@teklifpro.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "success": true,
                  "message": "Verification email sent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify email token",
        "description": "Verifies the email token sent in verification email. User clicks link in email which calls this endpoint. | E-posta doğrulama belirtecini doğrular.",
        "operationId": "verifyEmailToken",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "success": true,
                  "message": "Email verified successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from registration or login"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Invalid input parameters"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
    }
  }
}